Support POST for creating a customer
Update tests to test both GET and POST operation of the customers
endpoint. Also update database constraints to not allow the insertion of
empty text columns.

The JSON:API requires the "id" field to be a string, yet we base them on
integers from PostgreSQL, so the datatype in our struct will be int64
instead. This requires us to implement jsonapi.MarshalIdentifier to show
how the struct integer is converted to a string format.

The id also has another interesting issue: for server operation it is
required to include the id when generating JSON, but for clients
creating JSON to send to a server it is optional. Since we use the same
"customer" struct for both things we need to add the omitempty tag to
allow the server to unmarshal data from a client that has not set an id.

The other way around our tests also needs to use a MarshalClientMode
option to skip the id check when creating JSON data.