Convert id PRIMARY KEY columns from bigint to UUID
It is helpful to have globally unique identifiers for rows even across
multiple tables. It is less of an issue to confuse one "id 10" with
another tables "id 10" if they are all unique.

This has the added benefit that tests (since we set the UUIDs by hand
there instead of having the database create them for us) can refer to
id's that can only match one thing across all migration files.