Start storing session keys for cookies in db
When trying to add tests for server.Init() it became tricky to manage
the goose migration files across both the "migrations" package as well
as the testdata because all .go migration files would be loaded via
init() calls even if we only expected to load the ones for our
migrations package.

For this reason stop using migration files for filling in testdata and
only use it for managing structural changes to the database.