Add WIP web console
The console is what you interact with from a web browser. Currently
login is performed via the same database users as the API calls. The
idea is to have the login page also point to OIDC auth via keycloak.

Once console authentication is successful a cookie based session is
created which also means we need CSRF protection, this is handled via
gorilla/csrf.

The web pages themselves are generated using templ, so when modifying a
.templ file the code in the "components" module needs to be regenerated
and this can be done by running `go generate ./...`.

Also add --dev flag to "server" command. This disables the "secure" flag
for cookies used for sessions and CSRF handling so we can run locally
without TLS.