Add initial OIDC code
Adds "log in with keycloak" link on login page. Clicking it currently
only does the OIDC code flow dance and prints the resulting id token
contents. Next steps is to actually create a matching user in the local
db if login is successful etc.

Code is mostly based on the oidc example at
https://github.com/coreos/go-oidc/blob/v3/example/idtoken/app.go