Add concept of auth providers
Now that we are about to create users based on OIDC logins via keycloak
rather than only using local accounts with passwords we need to be able
to know what auth provider a given user was created from. For now call
the providers "local" and "keycloak".

The idea is to use this information to be able to look up provider
specific information in other tables, e.g. for keycloak users we
probably need to store the "sub" claim found in the OIDC id_token (this
being the "real" user id of the equivalent user created in keycloak, so
not bound to a username, email address or other things that
might change over time).