Add support for logging in Keycloak users
A successful login via keycloak now creates a local entry in the users
table. If the username is changed in keycloak we notice this and update
it in the database as well (this is possible becuase we track users via
the OIDC sub claim which is UUID that should never change (unless we
lose the keycloak database or manually remove a user).

Some additional fixes:
* Set global session cookie options on the cookistore
  rather than indvidual sessions as I noticed options are not persisted
  on Save().
* Rename default non-superuser role from "customer" to "user".