Add initial support for getting org IPs
Also include helper function to verify if user is allowed to carry out
the request.

Also the CHECK for network() and broadcast() on the org ip address
does not work since we force a /32 or /128 masklen where these would
all return the same value. Ideally we would check this against the
network CIDR the address is linked to, but not sure we want to go down
the path of creating some stored procedures to check this inside the
database for now, maybe it is easier to keep such cross-table checks in
the app code.