Replace PATCH with PUT for user modifcations
This is a more basic operation that requires the user to supply the full
user data in the request, but leaves the more complex details of "how
data sent in a PATCH request is actually handled", e.g. should it be
"JSON Patch" formatted? Should it be "JSON Merge Patch" formatted?" etc.

Also make both creation and update of users return a user struct for
returning to the API caller.