Add service version inspection possible in console
Add separate types pkg for sharing types between other packages, in
this case the server can fill in service versions info and pass it to
the console component.

While working on this the inconvenience of of using NULL in the
service version "active" column to signify "false" became apparent.
The reason for this "true is true, NULL is false" is that it means we
can use a simple UNIQUE constraint to only have a single row be "true"
while all others are NULL.

Now the UNIQUE constraint has been replaced with a "partial UNIQUE
index" instead, resulting in the UNIQUE index only looking at true
values. This allows us to guarantee only a single row can be true while
still allowing all others to be false.

This does require a more complex "CREATE UNIQUE INDEX" in the migration
file, but at least it becomes a lot more convenient to work with from
the application.

References:
https://dba.stackexchange.com/questions/197562/constraint-one-boolean-row-is-true-all-other-rows-false
https://www.postgresql.org/docs/current/indexes-partial.html