Run tests with pgConfig.MaxConns = 1
This way we ensure that functions can survive with only one connection
available. If tests hang this probably means you are trying to use a
dbPool inside an existing transaction. If this is the case then convert
things so you pass on the tx instead of the dbPool.

Make dbUserLogin() use tx to fix hanging test after the change:
The system works!