Add newUserIdentifier() and use it in selectUser()
Include additional orgID and roleID in userIdentifier type as this
made the original SELECT in selectUser() unnecessary.

Also add "FOR SHARE" on the different resource identifier queries as
this is the theme for all the transaction based helper functions we
have. Should result in less need to retry operations because of
concurrent changes.

selectUser() was renamed from selectUserByID() as the function takes
either ID or name as a parameter.