Add concept of service quota
Since a service is the entity that is bound to resources like varnish
instances and IP addresses make it possible to limit how many services a
given org can create, limit to 1 by default.

Set a large quota for "org1" in the tests to keep them working as usual
while creating a new "org4" that can be used to test the limit.