Add VCL validation
The service now validates that the VCL config being added either via API
or the console is actually valid. It does this by a HTTP POST to a
separate service (https://github.com/SUNET/sunet-vcl-validator) which is
expected to run separately.

To make this work during tests we start such a container using the
https://testcontainers.com framework.

Replace a bunch of hostnames in the VCL-testing-data with IP addresses
as the validation can get upset if it fails to resolve a DNS name.