Add API for fetching full node config
Available at /api/v1/cache-node-configs this is supposed to be used by
sunet-cdn-agent to collect the complete view of the world in order to
build local configuration as well as detect changes for what is already
present on disk.

While here rework generateCompleteServiceVcl() to take the specific
arguments needed instead of a whole ServiceVersionConfig struct and only
use parts of it. Makes it easier to call it at places where we do not
have access to a ServiceVersionConfig.