Allocate UID range to service
We want to supply a shared directory with configs used by
different containers for a service. This is done so we can atomically
switch between versions so all containers see the same config version
via symlink). The shared config dir does mean that potentially one
container can read the config for another container, and to be able to
control this we want to run the different containers with different
UIDs. For this to work we need a service to have access to a range of UIDs
rather than a single one.