Rework Varnish config generation
We need to be able to control what certificate name is expected when
setting up the outgoing TCP connection via HAProxy to an origin server,
and HAProxy is configured to get this name from the authority TLV in the
PROXY2 header that varnish creates. We can control this TLV with the
.authority setting in Varnish.

Since the .host and .port settings of the backend is used to set the
PROXY2 dst_addr/dst_port which shows "where the client originally
connected" we set these to the information of the frontend, not the
origin as we would do with a normal backend. For now we just use the
IPv4 address, it is possible the client connected to the IPv6 address
but ignore this for now.

While here update some database constraint names to always include the
column name they are relating to as this makes future expansion cleaner.

While working on this start using templating for the varnish config
instead of generating the config line by line. This makes it easier to
get an overview of the config for future modification and works the same
as the config generation used in sunet-cdn-agent.