Add domains table
Since the sunet-cdn-agent will assign certdata to containers based on
the domain name in the service version we need to make sure that a
domain is actually verified as owned by the organisation that supplied
the configuration.

Now when creating a service version with domain names in it we verify
that the domain exists in the domain tables and is verified.

Actual validation logic that looks for unverified domains and tries to
verify them is still needed.