Add DNS verification for domains
Basically we do a TXT query for any unverified domain in the database
and expect to find a record containing
"sunet-cdn-verification=<verification token>". If this is found we mark
it as verified which in turn allows the name to be used when submitting
domains to service versions.

The code expects to ask a resolver for the information rather
than recursing itself as this is much easier to implement. This does
mean that the cache of that resolver could get in the way of getting
the latest DNS changes. Ideally a resolver used by this code is configured
to limit all TTLs to a maximum of something low like 60 seconds to never
be affected by old data that could be lingering in the cache. Looking at
forum posts like
https://community.letsencrypt.org/t/letsencrypt-is-reading-an-outdated-dns-txt-record/102934
this is basically how lets encrypt does it.

Currently no "recheck" of an already verified domian is ever performed.