Add Domains page to console
This also resulted in reworking the domains API endpoint. We now support
not supplying an org and the API will then return either all domains if
the requesting user is a superuser or domains belonging to the given
user if a normal user.

Since calls no longer requires specifying an org the API endpoint also
moved from /v1/orgs/{org}/domains to /v1/domains with org supplied as an
optional query param.