Remove OrgName and ServiceName from node config
I have realised that it is better to only stick to IDs in files and
paths generated by sunet-cdn-agent on the cache nodes. Since IDs are
immutable it is one less thing that could change and result in things
having to reload on the cache nodes due to modified config files etc.

Another improvement with replacing names with IDs is that they are
generated by the database so is one less attack surface for potential
injection attempts.

Since sunet-cdn-agent no longer uses the names remove them from the data
we send, this way it is harder to accidentally use them somewhere in
future work.