Remove "expose fd-listeners"
As haproxy is running in "master-worker" mode (-W) this should no longer
be needed since at least 2.6, from
https://docs.haproxy.org/2.6/configuration.html#5.1-expose-fd%20listeners:
```
In master-worker mode, this is not required anymore, the listeners will be
passed using the internal socketpairs between the master and the workers.
```

Remove it to not get the impression that the socket must be enabled for
seamless reloads to work.