Rework varnish -> haproxy PROXYv2 config
Up to this point varnish was configured based on the "TLS onloader"
pattern as described here:
https://varnish-cache.org/docs/trunk/users-guide/vcl-backends.html#connecting-through-a-proxy

This has seemd to work OK, with the only caveat that the dst address is
statically set in the PROXY header based on the varnish ".host" config,
so always the IPv4 address, never the IPv6 address even if this is where
the client originally connected, because we had to set "one of them".

However, upon inspection of the haproxy "stick-table" state it contained
just one "src" entry:
===
$ echo show table haproxy_https | socat stdio unix-connect:/shared/admin.sock
0x7fc4f4691910: key=::ffff:0.0.0.0 use=0 exp=0 shard=0 server_id=1 server_key=s0
===

This was weird, as the haproxy config was set to "stick on src" I would
have expected it to contain the actual client source address, not some
empty-looking 0.0.0.0.

On closer inspection (reconfiguring varnish to communicate PROXYv2 with
haproxy via TCP instead of unix socket) and tcpdumping/wiresharking it
turned out that varnish would set the source IP in the PROXY header to
0.0.0.0 when using the above mentioned ".via" setup.

However, since we are already pointing out backend IP addresses in the
haproxy config instead of the magic "0.0.0.0:0" for picking it up from
PROXY header I have now replaced the ".via" setup in varnish with just a
basic backend pointing to haproxy with the addition of ".proxy_header":
https://varnish-cache.org/docs/trunk/reference/vcl-backend.html#attribute-proxy-header

When using this varnish instead seems to preserve the source and dest
addresses it received in the incoming PROXY header from haproxy and
include those in the new PROXY header it creates when connecting back to
haproxy. Now the stick-table is properly filled in with actual client IP
addresses (both IPv4 and IPv6) and the dest address in the PROXY header
sent from varnish to haproxy is what the client actually connected to
(IPv4 or IPv6).

Since we can no longer use .authority (only used with .via) to set the
SNI name for haproxy we move the SNI hostname config to haproxy as well.