Switch haproxy to http mode for all listeners
They were put in tcp mode based on the example at
https://varnish-cache.org/docs/trunk/users-guide/vcl-backends.html#connecting-through-a-proxy
but http mode seems to work as well and gives us better logs.

Since we are terminating TLS in haproxy there is probably no need to do
TCP mode. Given that we would like to do SNI based on req.hdr(host) as
strongly recommended in
https://docs.haproxy.org/2.9/configuration.html#sni
for requests to origin servers it is also helpful to actually have
access to HTTP headers.

Remove link to varnish docs since we are not really using the "TLS
onloader" pattern anymore, having replaced it with ".proxy_header".