Add support for cache node maintenance mode
Make it possible to manage mainteance mode of cache nodes by setting the
"maintenance" column in the database to true. The idea is that this will
control the ipvsadm weight in l4lb nodes. The default value is true
so that if you add a cache node without setting it explicitly the node
will not automatically be sent traffic right away.

The setting can be managed via the
/v1/cache-nodes/{cachenode}/maintenance endpoint.

While here add "name" column to cache nodes. Since most resources has
ID and name columns lets have this here as well. This makes the above
API endpoint behave like other resource types where both an ID and name
is valid. Note that the name follows the same rules as other "name"
columns of being a single DNS label, so if we want to store FQDNs in the
future this probably needs to go in a dedicated "hostname" column or
similar.