Some HTTPS handling rework
Drop support for setting a "SNI hostname" when connecting to an origin
server for now. It is fairly complicated to decide if we should support
rewriting either the HTTP Host header or the SNI hostname when
connecting to an origin, and if rewriting either of these should result
in also rewriting the other one to match. So skip this until we have a
real need for it.

Instead introduce a "Verify TLS" setting per origin server that decides
if we should do validation of the cert data received from the origin. If
this is the case we are strict and expect the certdata from the origin
to cover the HTTP host header that was originally requested by the web
client, and if it is disabled we do not care at all about the validity
of the cert data.

This way we support either "very strict" validation that should be the
preferred mode, or allow completely unverified connections if it is not
possible to use valid certs for some org for some reason.