Update service version creation in console
Split out port from the origin text field since it is easier to validate input
if the host and port are different fields (and also easier for a user to
understand what should be entered). Also keep each origin in its
own nested struct instead of bundling settings for multiple origins in
slices for each type.

This also requires the ".0." naming e.g. name="origins.0.port" in the
HTML data so gorilla/schema knows how to bundle things into structs.