Add DoS protection for argon2 hashing
Since argon2 hashing is designed to be resource intensive this also
becomes a source of DoS problems if many requests are handled
concurrently, like possibly killing the process due to OOM issues.

For this reason add a mutex that makes it so at most one request at a
time can call the resource intensive argon2 operation, and also
introduce a login cache so we do not need to redo hash calculations for
users that have already managed to login since we started.

Finally purge the login cache if any user changes their password just to
be sure the cache only contains up-to-date hashes.