Start handling console forms better
Starting with the domain creation form: we need to retain input values
so rendering the page again can be prefilled, and if there are
validation errors we need to display the error next to the field it
belongs to.

For this introduce structs that holds both the input values as well as
errors related to the same input. One needs to be careful when echoing input
data back to the user to not introduce XSS vulnerabilities but the templ
library takes care of HTML attribute encoding automatically:
https://templ.guide/syntax-and-usage/attributes/#string-expression-attributes

Also, since the "fqdn" validation tag allows a trailing "." for FQDNs
make sure we strip out such characters prior to inserting the name in
the database since we store domain names without the root label there.