Start using CSS Grid to arrange console content
This way we can have a header for things like the logout button, a sidebar
navigation field for selecting the different configuration pages and a
main box for displaying the contents of the currently selected page.

While here move out the CSS from the inline <style> tag to separate
.css files to make the templ component less bloated.

Also rename headerComponent to headComponent since it actually manages
the <head> tag, and we now also use the <header> tag for the grid
content.