Improve login page
Make js and css available directly under /js and /css to make them
usable outside of just the console pages.

Disable hx-boost for the logout link so we do a normal full page reload
when clicking it rather than doing htmx DOM-replacement where things
like currently loaded css is left in place. Seems reasonable for a
logout link to get a clean slate.

For now do not pull in htmx at the login page since it is not needed for
the basic form.