Add workaround for regression in gorilla/csrf
In v1.7.3 the CSRF protection will always reject plain HTTP requests
which breaks local development.

Add middleware that will call csrf.PlaintextHTTPRequest() for requests
if we are not running in secure mode. This signals to the CSRF
middleware that the request is being served over plaintext HTTP and that
Referer-based origin allow-listing checks should be skipped.