Add _hyperscript for client side scripting
Right now just used to make flash message handling a little prettier as
a proof of concept, but will be needed to make it possible to
dynamically extend the service version form with more than one origin.

Rearrange both js and css directories so external libraries like htmx
and _hyperscript lives in versioned directories. Makes it easer to know
what we are using for future maintenance.