Replace peer IP socket with unix socket
Because haproxy runs in the host network namespace 127.0.0.1 is shared
between all haproxy processes. This means spinning up another service
leads to it not being able to start:
```
[ALERT]    (1) : Binding [/service-versions/active/haproxy/haproxy.cfg:29] for peers section local_haproxy: cannot bind socket (Address already in use) for [127.0.0.1:1024]
[ALERT]    (1) : [haproxy.main()] Some protocols failed to start their listeners! Exiting.
```

One approach would be to start registering a port for each instance of
haproxy, or using another local address or something, but seems simpler
to just use a unix socket instead.