Add service version clone support to UI
Now there exists a button for each service version that opens the
service version create form but with all fields filled in based on the
version that is being cloned.

This should allow the general workflow of starting from a known good
config and then modyfing it based on needed changes.

Move some VCL step struct handling code to the types package so it
can be reused by the console UI code when looking up the cloned data for
each step.