Add more than one origin via web console
With some htmx magic it is now possible to click a button to add
additional origin fieldsets to the service version creation form.

What happens is that clicking the button will send an AJAX GET request
via hx-get to the new /console/new-origin-fieldset endpoint and this
will return the new <fieldset> HTML. One thing that was tricky is that
we need to keep track of the name counter used for the struct mapping
(e.g. origins.0.host, origins.1.host). To do this the initial form will
include a hidden <input> field with the value of the next available
index (e.g. if the initial page includes a single origin the next index
will be 1).

We then set the htmx hx-swap-oob attribute in the returned response
which allows us to replace the <input> field with the updated counter.
This is what allows us to click the button multiple times and each time
get an updated index counter in the data returned from the server.