Simplify "add origin" button
Instead of tracking the current index number in a separate <input> just
store it in the button itself. Now tmx will just replace the whole
button element on each click which is just simpler and also means we are
not including the contents of the hidden <input> element when POSTing
the form as well.