Enable dynamic lookups of origin servers
Now when adding origins with hostnames rather than IP addresses haproxy
will continously (based on "timeout resolve" interval, which we are
currently leaving at the default value of 1s) look up DNS records to
find the IP address of the origin.

Many online resources mention you may also want to configure "init-addr" to include
"none" so haproxy is able to start even if the hostname does not resolve,
but it seems that since commit
https://github.com/haproxy/haproxy/commit/ac1c60fd9c63dd9ac4bab7e9a9f24fb347a5fbdf

... "none" will be automatically added to "init-addr" config if using dns
resolvers, so just skip configuring an explicit "init-addr", we will
still be able to start.

Right now just use the contents of resolv.conf to decide what nameserver
to use, and increase accepted_payload_size to what seems to be generally
given as example in the manual.