Improve origin handling in service version console
Before this the first origin in the service version creation form did
not have a delete button on it, with the idea that it wasnt needed since
we expect at least one origin to be entered. However of someone adds
a second origin based on the first and want to then remove the first one
we should support that as well.

Now all origins have a delete button, but if there is only one origin
left we will disable the button. Since we use htmx swapping to add new
fieldsets we listen for the htmx:afterSwap in order to have the new
content prior to counting the how many we have.