Add support for removing users
Superusers are now able to remove users (except for themselves). This
also requires updating SQL tables referencing the users table to cascade
the delete if the user is removed.

Include tests that verify users exists and and are removed as expected
and include filling in referencing tables so we verify that ON DELETE
CASCADE is present.