Make it possible to get services by organization
Before this change the list of services would be either "all services"
if requested by a superuser or the list of services belonging to the
same org as the requesting user.

Being able to select the services from a given org is needed if a
superuser needs to manage services for that org. This is needed for
upcoming changes to the console.

Add tests for the API endpoint that uses this.

If no org is requested the old behaviour is used.