Move main console dashboard under org/<org-name>
Prior to this all operations were done directly under /console. This is
problematic if a superuser needs a way to select what organisation it is
working on.

Now the URL will containt /console/org/<org-name>/... instead, so
switching from one org to another is a matter of changing the URL.

More work is needed for the "root" console page at /console where a
superuser with no org configured will end up after login (like the main
admin account) but push the current WIP state for now.