Make org-selector handle no org relation
For the default "admin" superuser which does not belong to an
organization we need make the org selector UI display something when the
user is at the root /console path rather than an org URL.

Now it has the option "-- not selected --" which also allows you to navigate
back to /console after having switched to some /console/org/org-name URL.

The /org-switcher now knows about this special org name and redirects
to /console if that is what is sent. Since "-- not selected --" is not a
valid DNS label which is a requirement for real org names there should
be no risk of a collision with a real org.