Add support for "origin groups"
This makes it possible to add origins to either the always present
"default" group which gets some automatic handling in the generated VCL
rules, or you can add them to a custom origin group where you then need
to explicitly use them in the VCL code you write, e.g. in vcl_recv:
```
set req.backend_hint = my-origin-group_http;
set req.backend_hint = my-origin-group_https;
```

In the above case the name of the group is "my-origin-group" but it
contains at least one origin that enables TLS and one that does not so
you need to refer to them both in the VCL code for it to be accepted by
the varnish config validator.

What is currently missing is a way to create origin groups in the UI
console.