Make "init" command take password file as input
Instead of having the manager generate and output the password during
initializaion feed it via a file instead.

When running this system via docker compose it becomes cumbersome to run
"init" while keeping the outputted text from being logged (therefore
leaking the initial password wherever the logs are being sent).

This way the operator is responsible for choosing the password but it
does make it easier to not leak secrets no matter how it is run.