Add concept of node groups
Since we want to control what cache nodes a given l4lb sends traffic to
(e.g. to keep traffic inside the same site) we need a way to group nodes
togehter.

With this we introduce the concept of "node groups" which allows you to
control what cache nodes the l4lb will use as backends as well as giving
the cache nodes a way to open e.g. firewall rules from the l4lb nodes in
the same group.

This changes the /v1/cache-node-configs and /v1/l4lb-node-configs to
require supplying an additional path parameter
(/v1/cache-node-configs/my-node-name) so the manager can know from what
node the contained config should be suitable for.

Cache nodes have existed in the database already but this change also
brings in l4lb nodes since these are needed for the mentioned firewall
management on the cache nodes.

The returned data now also contains the l4lb or cache node contents for
the requested node name or ID so the node itself can tell if it is in
maintenance mode or not etc.