Support multple addresses for nodes
This is needed for l4lb nodes that has multiple paths to the internet
today but lets do it for cache nodes as well to keep things the same.

It is possible we should consolidate nodes into a single table to not
cause more duplicate code handling each type with separate tables.

The addreses now mix ipv4 and ipv6 addresses to not cause double amount
of address tables. This caused problems for huma regarding validation of
the []netip.Addr where it seems to expect a []netip.Addr to
automatically mean IPv4 only. Introduces a "IPAddress" type to work
around that problem.