Keep form data when service version submit fails
Until now failure to submit a service version form has lead to coming
back to the form with all fields empty. This is not very helpful when
working on e.g. a large VCL rule or similar.

Now the current contents of the form is passed back to the page generation
so we can include it when writing out the page again. The reason for
passing in the form struct via pointer is that the struct contains
boolean values and we need to be able to tell if a false bool value
should be considered set as opposed to just being a zero value.