Upgrade testcontainers-go
After adding the new testcontainer for keycloak running
`go test -race ./...` would fail like so (not that it is the already
existing test that failed, not our new one):
```
--- FAIL: TestPostServiceVersion (0.05s)
    server_test.go:3344: create container: reaper: container state: Error response from daemon: No such container: 57462ae8987bf6f9deff92b9bdc455cc1fac6de7075d4bb93af6381553e59fcd
        terminate: stop: container stop: Error response from daemon: No such container: 57462ae8987bf6f9deff92b9bdc455cc1fac6de7075d4bb93af6381553e59fcd
```

It seems the container ID it was looking for was the ryuk reaper container
that handles cleanup.

Running tests without -race did not show the problem.

Upgrading seems to fix it so -race works again:
```
go get github.com/testcontainers/testcontainers-go
go mod tidy
```